4.23

+++ part4/blogi/controllers/blogs.js
@@ -1,11 +1,14 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const jwt = require('jsonwebtoken')
+const User = require('../models/user')
 
-blogsRouter.get('/', (request, response) => {
-  Blog.find({}).then(blogs => {
-    response.json(blogs)
-  })
+blogsRouter.get('/', async (request, response) => {
+  const blogs = await Blog
+    .find({}).populate('user', {username: 1, name: 1, id: 1})
+  response.json(blogs)
 })
+
 blogsRouter.get('/:id', async (request, response) => {
   const blog = await Blog.findById(request.params.id)
   if (blog) {
@@ -14,27 +17,44 @@ blogsRouter.get('/:id', async (request, response) => {
     response.status(404).end()
   }
 })
-blogsRouter.post('/', (request, response) => {
+
+blogsRouter.post('/', async (request, response) => {
   const body = request.body
+  const user = request.user
+  
+  if (!user) {
+    return response.status(401).json({ error: 'token missing or invalid'})
+  }
+  if (body.title === undefined || body.url === undefined) {
+    return response.status(400).json({ error: 'missing title or url'})
+  } 
+  
   const blog = new Blog({
       title: body.title,
       author: body.author,
       url: body.url,
       likes: body.likes ? body.likes : 0,
+      user: user.id,
   })
-  if (body.title === undefined || body.url === undefined) {
-    return response.status(400).end()
-  } 
-  blog
-  .save()
-  .then(result => {
-    response.status(201).json(result)
-  })
+  
+  const newBlog = await blog.save()
+  user.blogs = user.blogs.concat(newBlog._id)
+  await user.save()
+  response.status(201).json(newBlog)
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  const user = request.user
+  if (!user) {
+    return response.status(401).json({ error: 'token missing or invalid'})
+  }
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(401).json({ error: 'unauthorized deletion'})
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
